Moved keys out of pkg
