+++ cmd/client/main/main_test.go
@@ -15,7 +15,7 @@ import (
 
 	"github.com/SIGBlockchain/project_aurum/internal/client/src/client"
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/hashing"
-	"github.com/SIGBlockchain/project_aurum/pkg/publickey"
+	"github.com/SIGBlockchain/project_aurum/internal/publickey"
 )
 
 var removeWallet = true
