+++ cmd/handlers/main.go
@@ -17,7 +17,7 @@ import (
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/blockchain"
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/contracts"
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/hashing"
-	"github.com/SIGBlockchain/project_aurum/pkg/publickey"
+	"github.com/SIGBlockchain/project_aurum/internal/publickey"
 
 	"github.com/SIGBlockchain/project_aurum/internal/endpoints"
 	"github.com/SIGBlockchain/project_aurum/internal/handlers"
