+++ internal/client/src/client/client.go
@@ -22,8 +22,8 @@ import (
 
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/accountinfo"
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/hashing"
-	"github.com/SIGBlockchain/project_aurum/pkg/privatekey"
-	"github.com/SIGBlockchain/project_aurum/pkg/publickey"
+	"github.com/SIGBlockchain/project_aurum/internal/privatekey"
+	"github.com/SIGBlockchain/project_aurum/internal/publickey"
 )
 
 var SecretBytes = hashing.New([]byte("aurum"))[8:16]
