+++ internal/producer/src/block/block_test.go
@@ -14,7 +14,7 @@ import (
 
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/contracts"
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/hashing"
-	"github.com/SIGBlockchain/project_aurum/pkg/publickey"
+	"github.com/SIGBlockchain/project_aurum/internal/publickey"
 	"github.com/google/go-cmp/cmp"
 )
