+++ internal/producer/src/contracts/contracts.go
@@ -11,7 +11,7 @@ import (
 	"reflect"
 
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/hashing"
-	"github.com/SIGBlockchain/project_aurum/pkg/publickey"
+	"github.com/SIGBlockchain/project_aurum/internal/publickey"
 )
 
 /*
