+++ internal/producer/src/contracts/contracts_test.go
@@ -13,7 +13,7 @@ import (
 	"testing"
 
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/hashing"
-	"github.com/SIGBlockchain/project_aurum/pkg/publickey"
+	"github.com/SIGBlockchain/project_aurum/internal/publickey"
 )
 
 func TestNew(t *testing.T) {
