+++ internal/producer/src/validation/validation.go
@@ -10,7 +10,7 @@ import (
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/accountstable"
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/contracts"
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/hashing"
-	"github.com/SIGBlockchain/project_aurum/pkg/publickey"
+	"github.com/SIGBlockchain/project_aurum/internal/publickey"
 )
 
 func ValidateContract(c *contracts.Contract) error {
